v1.15 - rollback, restore, MCM fix
Apparently some GAMMA players experienced a bug with 1.13 due to the Track monkeypatch. Should be impossible for my patch to affect their values, but that's Stalker for you.

I released 1.14 as a breakfix rollback to 1.12.2's code with an incremented version. After identifying the Track patch as the sole culprit, I then released 1.15, which includes all of 1.13 minus the Track-based alpha preview, and adds:
* Completed overhaul of pos preview menu, with full-size widget preview
* Fixes for markers not disappearing on NPC death, pins not reappearing on map after savegame load, and debug logging getting re-enabled