v1.13, 1.14, 1.15
The Track element monkeypatch in v1.13 caused some GAMMA users to experience an issue with MCM slider values. I immediately reverted to 1.12.2 code, and released this breakfix as 1.14.

Once the Track patch was identified as the culprit, I released 1.15 which included all of 1.13--minus the Track patch, and with the WIP improvements I'd been testing for the next version.

* Added optional reticle for targeting valid pin/waypoint objects.
* Makeover of the widget position menu with preview using a miniature mock-up of 16:9 and 21:9 screens.
* Configure pins to fade in or out based on distance.
* Significant improvements to HUD icon scaling and positioning.
* Crash fixes related to nil value checks and body looting.
* Overhaul of Widget menu with full-size preview of current theme.
* Fix for pins not being re-added to the map on savegame load (they're still there, just weren't visible).
* Fixes for pins not clearing on NPC death, debug logging re-enabling itself, and some MCM values being incorrect.